Extract thumbnail-related functionality in ThumbnailsHelper
